requirements.txt
streamlit==1.31.1
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.15.0
opencv-python==4.8.0.74
plotly==5.18.0
pandas==2.0.3
streamlit-image-select==0.6.0
streamlit-option-menu==0.3.2 